refactor to support PTHR-90 resolution
The PTHR-90 has a resolution of 0.01 degrees, not 0.1. PTHR90 overrides
the _ANGLE_MULTIPLIER to get the correct angle values. Since the PTHR90
doesn't support the special 999.9 coordinate for "no move", that
functionality has been removed; that feature wasn't that important since
the user will know the current pan-tilt coordinates anyway, so they can
pass the current coordinate if they don't want an axis to move.